fix unrelated test flake
